cutomer字段有值了
