+++ src/main/java/com/mindata/ecserver/main/repository/primary/EcCustomerRepository.java
@@ -1,6 +1,6 @@
 package com.mindata.ecserver.main.repository.primary;
 
-import com.mindata.ecserver.main.model.primary.EcCustomerOperation;
+import com.mindata.ecserver.main.model.primary.EcCustomer;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
@@ -9,7 +9,7 @@ import java.util.List;
 /**
  * @author wuweifeng wrote on 2017/10/25.
  */
-public interface CustomerOperationRepository extends JpaRepository<EcCustomerOperation, Long> {
+public interface EcCustomerRepository extends JpaRepository<EcCustomer, Long> {
     /**
      * 查找某个客户最新的状态
      *
@@ -19,5 +19,5 @@ public interface CustomerOperationRepository extends JpaRepository<EcCustomerOpe
      *         分页
      * @return 客户状态
      */
-    List<EcCustomerOperation> findByCrmIdOrderByOperateTimeDesc(Long crmId, Pageable pageable);
+    List<EcCustomer> findByCrmIdOrderByUpdateTimeDesc(Long crmId, Pageable pageable);
 }
