+++ src/main/java/com/mindata/ecserver/main/repository/secondary/PtUserRoleRepository.java
@@ -9,6 +9,13 @@ import java.util.List;
  * @author wuweifeng wrote on 2017/10/26.
  */
 public interface PtUserRoleRepository extends JpaRepository<PtUserRole, Long> {
+    /**
+     * 查询某用户的所有role
+     *
+     * @param userId
+     *         userId
+     * @return 集合
+     */
     List<PtUserRole> findByUserId(Long userId);
 
     PtUserRole findByUserIdAndRoleId(Long userId, Long roleId);
